:recycle: [Refactor] SendFriendRequestResponse 클래스명 변경
